added navbar icons and moved out firebase
