+++ package.json
@@ -19,7 +19,7 @@
     "@react-navigation/native-stack": "^6.5.0",
     "expo": "~44.0.0",
     "expo-app-loading": "~1.3.0",
-    "expo-font": "~10.0.4",
+    "expo-font": "^10.0.5",
     "expo-google-fonts": "^0.0.0",
     "expo-status-bar": "~1.2.0",
     "firebase": "^9.6.6",
@@ -27,9 +27,11 @@
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
     "react-native-elements": "^3.4.2",
+    "react-native-feather": "^1.1.2",
     "react-native-gesture-handler": "~2.1.0",
     "react-native-safe-area-context": "3.3.2",
     "react-native-screens": "~3.10.1",
+    "react-native-svg": "^12.1.1",
     "react-native-web": "0.17.1"
   },
   "devDependencies": {
